requirements.txt
astroid==2.2.5
awscli==1.16.146
botocore==1.12.136
Click==7.0
colorama==0.3.9
docutils==0.14
filetype==1.0.5
Flask==1.0.3
Flask-GoogleMaps==0.2.6
geographiclib==1.49
geopy==1.20.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
pyasn1==0.4.5
pylint==2.3.1
pyodbc==4.0.26
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.24.2
Werkzeug==0.15.4
wrapt==1.11.1
xlrd==1.2.0
